+++ compiler/lib-runtime-files/js_of_ocaml_compiler_runtime_files.ml
@@ -23,7 +23,6 @@ let runtime =
     ; backtrace
     ; bigarray
     ; bigstring
-    ; bigstring_cstruct
     ; compare
     ; fail
     ; format